Moved the manifold to the core of the library
