+++ include/core/manifolds.h
